moving code into seperate files; restoring existing feature and test
