+++ lib/student.rb
@@ -1,0 +1,17 @@
+module George
+  class Student
+    def initialize(attributes)
+      @id = attributes[:id]
+      @gwid = attributes[:gwid]
+      @net_id = attributes[:net_id]
+      @grad_level = attributes[:grad_level]
+      @first_name = attributes[:first_name]
+      @full_name = attributes[:full_name]
+      @last_name = attributes[:last_name]
+    end
+
+    def email_address
+      "#{net_id}@gwu.edu"
+    end
+  end
+end
