+++ script/generate_rosters.rb
@@ -1,0 +1,13 @@
+#
+# to run from the root directory: `ruby script/generate_rosters.rb`
+#
+
+# Generate all current course enrollment report.
+
+require_relative "../lib/george.rb"
+
+current_courses = George::Term.current.courses
+current_sections = current_courses.map{|course| course.sections }.flatten
+current_sections.each do |section|
+  section.generate_roster # generate enrollment report
+end
